fix: split oversized backend files to satisfy backend-size CI
Extract managed-runtime tests and observability helpers into focused files
so the PR head stays within the 1000-line backend-size limit.
