+++ pkg/config/config_validator_test.go
@@ -545,34 +545,6 @@ func TestRuleResourceDefinitions_RequiresProviderQuotaMetadata(t *testing.T) {
 	assertFindingExists(t, findings, "resource-provider-quota-model")
 }
 
-func TestCanonicalStructuralFindings_RejectsUnsupportedManagedRuntimeIdentity(t *testing.T) {
-	cfg := testBaseConfig()
-	cfg.Resources = []interfaces.ResourceConfig{{
-		Name:       "unknown-cache",
-		Type:       interfaces.ResourceTypeModel,
-		Capacity:   1,
-		Model:      "UNKNOWN_RUNTIME",
-		Backend:    "LLAMACPP",
-		LoadPolicy: "ON_DEMAND",
-	}}
-
-	findings := CanonicalStructuralFindings(cfg)
-	assertFindingExists(t, findings, factoryvalidation.CodeManagedRuntimeUnsupportedIdentity)
-}
-
-func TestCanonicalStructuralFindings_RejectsLocalWorkerWithoutModelResource(t *testing.T) {
-	cfg := testBaseConfig()
-	cfg.Workers = []interfaces.WorkerConfig{{
-		Name:          "voice-local",
-		Type:          interfaces.WorkerTypeModel,
-		Model:         "OMNIVOICE_Q4_K_M",
-		ModelLocality: interfaces.ModelLocalityLocal,
-	}}
-
-	findings := CanonicalStructuralFindings(cfg)
-	assertFindingExists(t, findings, factoryvalidation.CodeManagedRuntimeWorkerMissingDep)
-}
-
 func TestRuleResourceDefinitions_AcceptsModelResourceMetadata(t *testing.T) {
 	cfg := testBaseConfig()
 	cfg.Resources = []interfaces.ResourceConfig{{
