+++ pkg/service/factory.go
@@ -51,26 +51,6 @@ type APIServerStarter func(ctx context.Context, runtime apisurface.APISurface, p
 
 type secretResolver = hostedworkers.SecretResolver
 
-// InvocationMetric records one emitted invocation counter together with its
-// low-cardinality dimensions.
-type InvocationMetric struct {
-	Name   string
-	Labels map[string]string
-}
-
-// InvocationMetricsRecorder receives invocation counter emissions from CLI and
-// session-runtime boundaries. Implementations should treat each call as a
-// single counter increment.
-type InvocationMetricsRecorder interface {
-	RecordInvocationMetric(InvocationMetric)
-}
-
-// ModelPullMetricsRecorder receives managed-runtime pull counter emissions.
-// Implementations should treat each call as a single counter increment.
-type ModelPullMetricsRecorder interface {
-	RecordModelPullMetric(InvocationMetric)
-}
-
 // ErrFactoryActivationRequiresIdle reports that runtime replacement was
 // attempted while the current runtime still had active work.
 var ErrFactoryActivationRequiresIdle = apisurface.ErrFactoryActivationRequiresIdle
