+++ pkg/service/model_event_classification.go
@@ -1,0 +1,36 @@
+package service
+
+import (
+	"errors"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/apisurface"
+	"github.com/portpowered/infinite-you/pkg/interfaces"
+	workerprovider "github.com/portpowered/infinite-you/pkg/workers/provider"
+)
+
+func modelEventDiagnostics(success *interfaces.WorkDiagnostics, err error) *factoryapi.SafeWorkDiagnostics {
+	if success != nil {
+		return interfaces.GeneratedSafeWorkDiagnosticsFromWorkDiagnostics(success)
+	}
+	var providerErr *workerprovider.ProviderError
+	if errors.As(err, &providerErr) {
+		return interfaces.GeneratedSafeWorkDiagnosticsFromWorkDiagnostics(providerErr.Diagnostics)
+	}
+	return nil
+}
+
+func modelEventErrorClass(err error) string {
+	var readinessErr *apisurface.ManagedRuntimeInvocationError
+	if errors.As(err, &readinessErr) && readinessErr.ReadinessState != "" {
+		return "MANAGED_RUNTIME_" + string(readinessErr.ReadinessState)
+	}
+	var providerErr *workerprovider.ProviderError
+	if errors.As(err, &providerErr) && providerErr.Type != "" {
+		return string(providerErr.Type)
+	}
+	if err == nil {
+		return ""
+	}
+	return "MODEL_EXECUTION_FAILED"
+}
